pyproject.toml
[project]
name = "act-ml-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azureml-core>=1.59.0.post1",
    # "azureml-sdk==1.59.0",
    "ipykernel>=6.29.5",
    "kagglehub>=0.3.10",
    "matplotlib>=3.10.1",
    "numpy>=1.23.5",
    "pandas",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "setuptools>=78.1.0",
    "shap>=0.47.1",
    "xgboost>=3.0.0",
]
